requirements.txt
# === Core ===
python-dotenv
requests
pymupdf           # for fitz (PDF text extraction)
sentence-transformers
pinecone-client

# === LangChain Splitter ===
langchain-text-splitters

# === Google Gemini API ===
google-generativeai

# === Optional / Utils ===
tqdm              # progress bars if you want
